Move simple exchange module
